Buffer the stageDone channel to prevent hanging goroutine
If the statemachine is closed and there is no more reader for the stageDone channel, then the `run` goroutine can hang waiting to write to stageDone.  Buffering the channel prevents this by allowing the chanable to writable even if the reader is gone.

The stageDone channel is now local to the `run` function as it was never used outside of it.